% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ora.R
\name{OraRun}
\alias{OraRun}
\title{Runs a SQL query in preferred environment.}
\usage{
OraRun(query, env = "dev")
}
\arguments{
\item{query}{SQL Query to execute.}

\item{env}{Environment name as string. The function will parse the environment name to its equivalent connection string stored under ora_connstr_[env]. This connection must be defined in advance that must contain five variables host_name, port, sid, user_name, pwd. Defaults to environment dev.
E.g., ora_connstr_dev <- list(host_name = "localhost", port = "1521", sid = "xe", user_name = "sc", pwd = "tiger")}
}
\value{
SQL resultset as a tibble
}
\description{
Runs a SQL query in Oracle and returns a resultset.
}
\examples{
ora_connstr_dev <- list(host_name = "localhost", port = "1521", sid = "xe", user_name = "sc", pwd = "tiger")
OraRun("SELECT * FROM ALL_TABLES", "dev")

}
