% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconresults.R
\name{ReconResults}
\alias{ReconResults}
\title{Compares two SQL resultsets from two database environments}
\usage{
ReconResults(query, env1, env2)
}
\arguments{
\item{query}{SQL Query.}

\item{env1}{Environment string to the first database.}

\item{env2}{Environment string to the second database.}
}
\value{
Returns a tibble of differences
}
\description{
Takes a SQL statement as an input and executes the SQL in two database environments, then compares the SQL resultset for differences.
}
